requirements.txt
django==2.2.10
djangorestframework==3.12.4
coreapi==2.3.3
pyyaml==6.0
